refactor: Create dedicated folder for artists api
