+++ project/api/artists/models.py
