+++ project/api/ping.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
-from flask_restplus import Api, Resource
+from flask_restx import Api, Resource
 
 ping_blueprint = Blueprint("ping", __name__)
 api = Api(ping_blueprint)
