+++ requirements.txt
@@ -1,6 +1,6 @@
 Flask==1.1.1
 Flask-Admin==1.5.5
-Flask-RESTPlus==0.13.0
+Flask-RESTX==0.1.1
 Werkzeug==0.16.1
 Flask-SQLAlchemy==2.4.1
 psycopg2-binary==2.8.4
